Add ESim sync functionality and enhance data models
Updated `ESimPackageList.razor` to include a sync button that triggers the new `SyncPackages` method. Removed the `commander` parameter from `AiraloController.cs` and deleted the `SyncAsync` method. Renamed methods in `EsimController.cs` for clarity.

Introduced new methods in `ESimOrderService.cs` for retrieving ESim orders and details. Added JSON serialization options in `ApplicationDbContext.cs` for `ESimPackageEntity`.

Enhanced `ESimPackageService.cs` with a `SyncPackages` method for Airalo API synchronization. Updated data types in `OrderPackageStatusView.cs` and added new properties in `ESimOrderEntity.cs`.

Introduced new classes for managing Airalo balance and updated views for ESim data. Updated project templates in `.vstemplates`. Overall, these changes improve the ESim management system's functionality and user experience.

+++ Shared/Features/AirAlo/Package/Views/OrderPackageStatusView.cs
@@ -16,7 +16,7 @@ public partial class OrderPackageStatusView
 public partial class OrderPackageStatusData
 {
     [DataMember, JsonProperty("remaining")]
-    public int Remaining { get; set; }
+    public double Remaining { get; set; }
 
     [DataMember, JsonProperty("total")]
     public int Total { get; set; }
